mlflow-client/requirements.txt
torch==2.6.0
mlflow==3.2.0
scikit-learn==1.7.1
numpy==2.2.2
pandas==2.2.2
ipykernel==6.30.1
pymysql
python-dotenv==1.1.1
boto3